@@ -1,6 +1,6 @@
 {
     "name": "obsidian-git",
-    "version": "2.30.1",
+    "version": "2.31.1",
     "description": "Integrate Git version control with automatic backup and other advanced features in Obsidian.md",
     "main": "main.js",
     "scripts": {
@@ -24,9 +24,7 @@
         "pnpm": ">=9"
     },
     "devDependencies": {
-        "@codemirror/state": "^6.4.1",
-        "@codemirror/view": "^6.34.1",
-        "@eslint/js": "^9.12.0",
+        "@eslint/js": "^9.14.0",
         "@types/debug": "^4.1.12",
         "@types/deep-equal": "^1.0.4",
         "@types/diff": "^5.2.3",
@@ -35,21 +33,26 @@
         "esbuild": "^0.24.0",
         "esbuild-svelte": "^0.8.2",
         "eslint": "^9.12.0",
-        "eslint-plugin-svelte": "^2.44.1",
+        "eslint-plugin-svelte": "^2.45.1",
         "obsidian": "^1.7.2",
         "prettier": "3.3.2",
         "prettier-plugin-svelte": "^3.2.7",
         "scss": "^0.2.4",
         "standard-version": "^9.5.0",
-        "svelte": "^4.2.19",
-        "svelte-check": "^4.0.4",
+        "svelte-check": "^4.0.5",
         "svelte-preprocess": "^6.0.3",
         "tslib": "^2.7.0",
         "typescript": "^5.6.3",
         "typescript-eslint": "^8.8.1"
     },
     "dependencies": {
+        "@codemirror/commands": "^6.7.1",
+        "@codemirror/merge": "^6.7.5",
+        "@codemirror/search": "^6.5.7",
+        "@codemirror/state": "^6.4.1",
+        "@codemirror/view": "^6.34.2",
         "buffer": "^6.0.3",
+        "codemirror": "^6.0.1",
         "css-color-converter": "^2.0.0",
         "debug": "^4.3.7",
         "deep-equal": "^2.2.3",
@@ -59,7 +62,8 @@
         "js-sha256": "^0.9.0",
         "obsidian-community-lib": "github:Vinzent03/obsidian-community-lib#upgrade",
         "simple-git": "^3.27.0",
-        "supports-color": "^9.4.0"
+        "supports-color": "^9.4.0",
+        "svelte": "^5.0.0"
     },
     "moduleFileExtensions": [
         "js",
