@@ -1,24 +1,43 @@
 <script lang="ts">
     import { setIcon, TFile } from "obsidian";
-    import { DIFF_VIEW_CONFIG } from "src/constants";
     import type { DiffFile } from "src/types";
-    import { getDisplayPath, getNewLeaf, mayTriggerFileMenu } from "src/utils";
+    import {
+        fileIsBinary,
+        getDisplayPath,
+        getNewLeaf,
+        mayTriggerFileMenu,
+    } from "src/utils";
     import type HistoryView from "../historyView";
 
-    export let diff: DiffFile;
-    export let view: HistoryView;
-    let buttons: HTMLElement[] = [];
+    interface Props {
+        diff: DiffFile;
+        view: HistoryView;
+    }
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access
-    $: side = (view.leaf.getRoot() as any).side == "left" ? "right" : "left";
+    let { diff, view }: Props = $props();
+    let buttons: HTMLElement[] = $state([]);
 
-    window.setTimeout(
-        () => buttons.forEach((b) => setIcon(b, b.getAttr("data-icon")!)),
-        0
+    let side = $derived(
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-member-access
+        (view.leaf.getRoot() as any).side == "left" ? "right" : "left"
     );
 
+    $effect(() => {
+        for (const b of buttons) if (b) setIcon(b, b.getAttr("data-icon")!);
+    });
+
+    function mainClick(event: MouseEvent) {
+        event.stopPropagation();
+        if (fileIsBinary(diff.path)) {
+            open(event);
+        } else {
+            showDiff(event);
+        }
+    }
+
     function open(event: MouseEvent) {
-        const file = view.app.vault.getAbstractFileByPath(diff.vault_path);
+        event.stopPropagation();
+        const file = view.app.vault.getAbstractFileByPath(diff.vaultPath);
 
         if (file instanceof TFile) {
             getNewLeaf(view.app, event)
@@ -28,59 +47,55 @@
     }
 
     function showDiff(event: MouseEvent) {
-        void getNewLeaf(view.app, event)?.setViewState({
-            type: DIFF_VIEW_CONFIG.type,
-            active: true,
-            state: {
-                file: diff.path,
-                staged: false,
-                hash: diff.hash,
-            },
+        view.plugin.tools.openDiff({
+            event,
+            aFile: diff.fromPath ?? diff.path,
+            aRef: `${diff.hash}^`,
+            bFile: diff.path,
+            bRef: diff.hash,
         });
     }
 </script>
 
-<!-- svelte-ignore a11y-click-events-have-key-events -->
-<!-- svelte-ignore a11y-no-static-element-interactions -->
-<!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<!-- svelte-ignore a11y_no_static_element_interactions -->
+<!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
 <main
-    on:click|stopPropagation={showDiff}
-    on:auxclick|stopPropagation={(event) => {
+    onclick={mainClick}
+    onauxclick={(event) => {
+        event.stopPropagation();
         if (event.button == 2)
             mayTriggerFileMenu(
                 view.app,
                 event,
-                diff.vault_path,
+                diff.vaultPath,
                 view.leaf,
                 "git-history"
             );
-        else showDiff(event);
+        else mainClick(event);
     }}
-    on:focus
     class="tree-item nav-file"
 >
     <div
         class="tree-item-self is-clickable nav-file-title"
-        class:is-active={view.plugin.lastDiffViewState?.file ==
-            diff.vault_path && view.plugin.lastDiffViewState?.hash}
-        data-path={diff.vault_path}
+        data-path={diff.vaultPath}
         data-tooltip-position={side}
-        aria-label={diff.vault_path}
+        aria-label={diff.vaultPath}
     >
         <div class="tree-item-inner nav-file-title-content">
-            {getDisplayPath(diff.vault_path)}
+            {getDisplayPath(diff.vaultPath)}
         </div>
         <div class="git-tools">
             <div class="buttons">
-                {#if view.app.vault.getAbstractFileByPath(diff.vault_path) instanceof TFile}
+                {#if view.app.vault.getAbstractFileByPath(diff.vaultPath) instanceof TFile}
                     <div
                         data-icon="go-to-file"
                         aria-label="Open File"
                         bind:this={buttons[0]}
-                        on:auxclick|stopPropagation={open}
-                        on:click|stopPropagation={open}
+                        onauxclick={open}
+                        onclick={open}
                         class="clickable-icon"
-                    />
+                    ></div>
                 {/if}
             </div>
             <span class="type" data-type={diff.status}>{diff.status}</span>
