@@ -1,9 +1,13 @@
 <script lang="ts">
-    export let files: unknown[];
+    interface Props {
+        files: unknown[];
+    }
+
+    let { files }: Props = $props();
 </script>
 
-<!-- svelte-ignore a11y-click-events-have-key-events -->
-<!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
 <main>
     {#if files.length > 500}
         <div class="tree-item nav-file">
