@@ -55,6 +55,7 @@ export interface ObsidianGitSettings {
     showFileMenu: boolean;
     authorInHistoryView: ShowAuthorInHistoryView;
     dateInHistoryView: boolean;
+    diffStyle: "git_unified" | "split";
 }
 
 /**
@@ -192,7 +193,7 @@ export interface Blame {
  */
 export interface FileStatusResult {
     path: string;
-    vault_path: string;
+    vaultPath: string;
     from?: string;
 
     // First digit of the status code of the file, e.g. 'M' = modified.
@@ -201,15 +202,12 @@ export interface FileStatusResult {
     index: string;
     // Second digit of the status code of the file. Represents status of the working directory
     // if no merge conflicts, otherwise represents status of other side of a merge.
-    working_dir: string;
+    workingDir: string;
 }
 
 export interface PluginState {
     offlineMode: boolean;
     gitAction: CurrentGitAction;
-    /* Currently refreshing the cached status
-     */
-    loading: boolean;
 }
 
 export enum CurrentGitAction {
@@ -241,10 +239,12 @@ export interface DiffEntry {
 
 export interface DiffFile {
     path: string;
-    vault_path: string;
+    vaultPath: string;
+    fromPath?: string;
+    fromVaultPath?: string;
     hash: string;
     status: string;
-    binary: boolean;
+    binary?: boolean;
 }
 
 export interface WalkDifference {
@@ -253,7 +253,7 @@ export interface WalkDifference {
 }
 
 export interface UnstagedFile {
-    filepath: string;
+    path: string;
     deleted: boolean;
 }
 
@@ -278,9 +278,29 @@ export type StatusRootTreeItem = RootTreeItem<FileStatusResult>;
 export type HistoryRootTreeItem = RootTreeItem<DiffFile>;
 
 export interface DiffViewState {
-    staged: boolean;
-    file: string;
-    hash?: string;
+    /**
+     * The repo relative file path for a.
+     * For diffing a renamed file, this is the old path.
+     */
+    aFile: string;
+
+    /**
+     * The git ref to specify which state of that file should be shown.
+     * An empty string refers to the index version of a file, so you have to specifically check against undefined.
+     */
+    aRef: string;
+
+    /**
+     * The repo relative file path for b.
+     */
+    bFile: string;
+
+    /**
+     * The git ref to specify which state of that file should be shown.
+     * An empty string refers to the index version of a file, so you have to specifically check against undefined.
+     * `undefined` stands for the workign tree version.
+     */
+    bRef?: string;
 }
 
 export enum FileType {
@@ -304,27 +324,72 @@ declare module "obsidian" {
     }
     interface View {
         titleEl: HTMLElement;
+        inlineTitleEl: HTMLElement;
     }
     interface Workspace {
+        /**
+         * Emitted when some git action has been completed and plugin has been refreshed
+         */
+        on(
+            name: "obsidian-git:refreshed",
+            callback: () => void,
+            ctx?: unknown
+        ): EventRef;
+        /**
+         * Emitted when some git action has been completed and the plugin should refresh
+         */
         on(
             name: "obsidian-git:refresh",
             callback: () => void,
             ctx?: unknown
         ): EventRef;
+        /**
+         * Emitted when the plugin is currently loading a new cached status.
+         */
         on(
-            name: "obsidian-git:view-refresh",
+            name: "obsidian-git:loading-status",
             callback: () => void,
             ctx?: unknown
         ): EventRef;
+        /**
+         * Emitted when the HEAD changed.
+         */
         on(
             name: "obsidian-git:head-change",
             callback: () => void,
             ctx?: unknown
         ): EventRef;
+        /**
+         * Emitted when a new cached status is available.
+         */
+        on(
+            name: "obsidian-git:status-changed",
+            callback: (status: Status) => void,
+            ctx?: unknown
+        ): EventRef;
 
+        on(
+            name: "obsidian-git:menu",
+            callback: (
+                menu: Menu,
+                path: string,
+                source: string,
+                leaf?: WorkspaceLeaf
+            ) => unknown,
+            ctx?: unknown
+        ): EventRef;
         trigger(name: string, ...data: unknown[]): void;
+        trigger(name: "obsidian-git:refreshed"): void;
         trigger(name: "obsidian-git:refresh"): void;
-        trigger(name: "obsidian-git:view-refresh"): void;
+        trigger(name: "obsidian-git:loading-status"): void;
         trigger(name: "obsidian-git:head-change"): void;
+        trigger(name: "obsidian-git:status-changed", status: Status): void;
+        trigger(
+            name: "obsidian-git:menu",
+            menu: Menu,
+            path: string,
+            source: string,
+            leaf?: WorkspaceLeaf
+        ): void;
     }
 }
