@@ -21,13 +21,7 @@ export default class DiffView extends ItemView {
         this.parser = new DOMParser();
         this.navigation = true;
         this.gitRefreshRef = this.app.workspace.on(
-            "obsidian-git:refresh",
-            () => {
-                this.refresh().catch(console.error);
-            }
-        );
-        this.gitViewRefreshRef = this.app.workspace.on(
-            "obsidian-git:view-refresh",
+            "obsidian-git:status-changed",
             () => {
                 this.refresh().catch(console.error);
             }
@@ -39,11 +33,11 @@ export default class DiffView extends ItemView {
     }
 
     getDisplayText(): string {
-        if (this.state?.file != null) {
-            let fileName = this.state.file.split("/").last();
+        if (this.state?.bFile != null) {
+            let fileName = this.state.bFile.split("/").last();
             if (fileName?.endsWith(".md")) fileName = fileName.slice(0, -3);
 
-            return DIFF_VIEW_CONFIG.name + ` (${fileName})`;
+            return `Diff: ${fileName}`;
         }
         return DIFF_VIEW_CONFIG.name;
     }
@@ -79,37 +73,37 @@ export default class DiffView extends ItemView {
     }
 
     async refresh(): Promise<void> {
-        if (this.state?.file && !this.gettingDiff && this.plugin.gitManager) {
+        if (this.state?.bFile && !this.gettingDiff && this.plugin.gitManager) {
             this.gettingDiff = true;
             try {
                 let diff = await this.plugin.gitManager.getDiffString(
-                    this.state.file,
-                    this.state.staged,
-                    this.state.hash
+                    this.state.bFile,
+                    this.state.aRef == "HEAD",
+                    this.state.bRef
                 );
                 this.contentEl.empty();
 
                 const vaultPath = this.plugin.gitManager.getRelativeVaultPath(
-                    this.state.file
+                    this.state.bFile
                 );
                 if (!diff) {
                     if (
                         this.plugin.gitManager instanceof SimpleGit &&
                         (await this.plugin.gitManager.isTracked(
-                            this.state.file
+                            this.state.bFile
                         ))
                     ) {
                         // File is tracked but no changes
                         diff = [
-                            `--- ${this.state.file}`,
-                            `+++ ${this.state.file}`,
+                            `--- ${this.state.aFile}`,
+                            `+++ ${this.state.bFile}`,
                             "",
                         ].join("\n");
                     } else if (await this.app.vault.adapter.exists(vaultPath)) {
                         const content =
                             await this.app.vault.adapter.read(vaultPath);
                         const header = `--- /dev/null
-+++ ${this.state.file}
++++ ${this.state.bFile}
 @@ -0,0 +1,${content.split("\n").length} @@`;
 
                         diff = [
@@ -134,7 +128,7 @@ export default class DiffView extends ItemView {
                     });
                     div.createEl("br");
                     div.createSpan({
-                        text: "File not found: " + this.state.file,
+                        text: "File not found: " + this.state.bFile,
                     });
                 }
             } finally {
